Spring Security: corrections for inserting in databasedefault department, admin- and user- roles and default employee for admin

+++ src/test/resources/delete_emp.sql
@@ -1,4 +1,4 @@
-delete from departments;
-delete from roles;
+delete from departments where depart_name <> 'default';
+delete from roles where role not in ('ROLE_ADMIN','ROLE_USER');
 --delete from users where login <> '$$$admin_login$$$';
 delete from users where login <> 'admin';
